fix: correct TestMockFileSystem to properly test file renaming
This commit fixes a bug in the mock filesystem test where the Rename
operation was not being called between creating a file and checking if
it had been properly moved. The test was writing to "/test/old.txt" and
then immediately checking if the file was gone and if "/test/dir1/new.txt"
existed, without actually calling fs.Rename().

The fix adds the missing call to fs.Rename() which properly tests the
renaming functionality of the mock filesystem implementation.

Additionally, the error checks were improved to provide better failure
messages in case of test failures.